Add documentation for uarchComputer
I added some documenation and also moved the tests out of the class they
were in in the testing suite. There still are some ambiguities about how
this operates since there is no microcode examples written, which also
means the main class is poorly covered in the testing infrastructure.
