+++ core/src/main/kotlin/org/eln2/compute/uarchComputer/FPU.kt
@@ -5,6 +5,9 @@ import kotlin.math.exp
 import kotlin.math.ln
 import kotlin.math.sqrt
 
+/**
+ * Floating Point Unit for uarchComputer
+ */
 @ExperimentalUnsignedTypes
 class FPU {
 	var inputA = 0.0
@@ -50,9 +53,15 @@ class FPU {
 			)
 		}
 
+	/**
+	 * FPU Operations List
+	 */
 	enum class FPUOps { Add, Cos, Div, Exp, Inv, Int, Ln, MinusOne, Mult, Neg, One, Sqre, Sqrt, Sub, ToDouble, Zero }
 	var currOp = FPUOps.Zero
 
+	/**
+	 * step - does computations on the registers. Load your data in before this, and grab your data after this.
+	 */
 	fun step() {
 		when (currOp) {
 			FPUOps.Add -> inputA += inputB
